requirements.txt
fastapi
starlette
gradio>=4.0.0
python-dotenv
langgraph
langgraph-cli[inmem]
langchain
langchain-core
langchain-openai
langsmith
authlib
uvicorn
itsdangerous
O365
google-auth-oauthlib
google-api-python-client
SpeechRecognition
email-validator
deepeval>=1.5.0
pytest>=8.0.0
